requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
aiofiles>=23.0.0
redis>=5.0.0
urllib3>=2.0.0
playwright>=1.40.0
asyncio-throttle>=1.0.0
psutil>=5.9.0
Pillow>=10.0.0